+++ crates/miden-tx/src/host/account_delta_tracker.rs
@@ -1,7 +1,10 @@
 use miden_objects::{
     Felt, ZERO,
-    account::{AccountDelta, AccountId, AccountStorageDelta, AccountVaultDelta, PartialAccount},
+    account::{AccountDelta, AccountId, AccountStorageHeader, AccountVaultDelta},
 };
+
+use crate::host::storage_delta_tracker::StorageDeltaTracker;
+
 // ACCOUNT DELTA TRACKER
 // ================================================================================================
 
@@ -14,29 +17,23 @@ use miden_objects::{
 ///
 /// TODO: implement tracking of:
 /// - account code changes.
-#[derive(Debug, Clone, PartialEq, Eq)]
+#[derive(Debug, Clone)]
 pub struct AccountDeltaTracker {
     account_id: AccountId,
-    storage: AccountStorageDelta,
+    storage: StorageDeltaTracker,
     vault: AccountVaultDelta,
     nonce_increment: Felt,
 }
 
 impl AccountDeltaTracker {
     /// Returns a new [AccountDeltaTracker] instantiated for the specified account.
-    pub fn new(account: &PartialAccount) -> Self {
+    pub fn new(account_id: AccountId, storage_header: AccountStorageHeader) -> Self {
         Self {
-            account_id: account.id(),
-            storage: AccountStorageDelta::new(),
+            account_id,
+            storage: StorageDeltaTracker::new(storage_header),
             vault: AccountVaultDelta::default(),
             nonce_increment: ZERO,
         }
-    }
-
-    /// Consumes `self` and returns the resulting [AccountDelta].
-    pub fn into_delta(self) -> AccountDelta {
-        AccountDelta::new(self.account_id, self.storage, self.vault, self.nonce_increment)
-            .expect("account delta created in delta tracker should be valid")
     }
 
     /// Tracks nonce delta.
@@ -49,8 +46,23 @@ impl AccountDeltaTracker {
         &mut self.vault
     }
 
-    /// Get a mutable reference to the current storage delta
-    pub fn storage_delta(&mut self) -> &mut AccountStorageDelta {
+    /// Returns a mutable reference to the current storage delta tracker.
+    pub fn storage(&mut self) -> &mut StorageDeltaTracker {
         &mut self.storage
+    }
+
+    /// Consumes `self` and returns the resulting [AccountDelta].
+    ///
+    /// Normalizes the delta by removing entries for storage slots where the initial and new
+    /// value are equal.
+    pub fn into_delta(self) -> AccountDelta {
+        let account_id = self.account_id;
+        let nonce_increment = self.nonce_increment;
+
+        let storage_delta = self.storage.into_delta();
+        let vault_delta = self.vault;
+
+        AccountDelta::new(account_id, storage_delta, vault_delta, nonce_increment)
+            .expect("account delta created in delta tracker should be valid")
     }
 }